Articulation: Clean up struct, widen EG2->Key mod
PanWidth, PanToVol, EG2ToPan, and LFOToPan were removed as they can't legally be defined by the DLS spec. I could add a metadata-driven option for them, but it's unlikely to be useful. EG2->Key was also widened to support the full range of keys, as needed by a port I'm working on.