Add sample-granularity key-on timing
This /massively/ helps time quantization issues, as the note onset is critical for beat perception. Because of the more precise timing, BPM resolution can be set as high as possible without adding quantization artifacts, as well as allowing lower update rates.

Also added a portamento bug fix for the edge case of a new note immediately after the last one ends.